requirements.txt
opencv-python
numpy
matplotlib
pillow
python-dotenv
streamlit
num2words
openai
gtts
tiktoken